+++ test/fixtures/basic-string-css-modules.js
@@ -1,0 +1,5 @@
+import css, { someClass } from './simple.css';
+
+__export__ = { css, locals: { someClass } };
+
+export default css;